Move Ranged definition to Fixer to avoid overpromising
